fix(headless): splitting actions in user session logic fixed (#4377)
## [SFINT-5703](https://coveord.atlassian.net/browse/SFINT-5703)


There was an issue with the logic that splits actions into multiple user
session, the function that does this is called
`splitActionsIntoTimelineSessions`.

This function takes an array of actions and a ticket creation date. The
actions received are sorted from the most recent to the least recent.
This function tries to regroup the actions that occurred in the same
session, Action A is considered in the same session as Action B if the
difference between the timestamps of the two actions is less than 30min.
The function does the following:

1- create a session object and initialize its start date and its end
date with the timestamp of the first action.
2- Loop on all the actions provided, for each action check if the
timestamp of the action occurred in less than 30min of the session end
date, if it is the case add the action to session and update the session
start date to be the timestamp of the newly added action.


### The issue:
In step 2, the issue is that we were comparing the time stamp of an
action to the session end date instead of the session start date.

Since the actions are sorted from the most recent to the least recent,
after adding an action to a session it's the session start date that is
changing thus this date is the one that needs to be used to know if the
next action will belong to the same session or not.



### Test
I updated the tests by updating the fake actions that are used in the
unit tests in order make sure we are no longer having the same issue and
that `splitActionsIntoTimelineSessions` behave as expected.


[SFINT-5703]:
https://coveord.atlassian.net/browse/SFINT-5703?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ