fix(atomic): fix commerce facets being re-attached to the DOM (#4372)
In commerce facets, we were missing defensive code around
unsubscribe/resubscribe VS connectedCallback/disconnectedCallback (ie:
stencil lifecycle when a facet gets moved around in the DOM).

End result is that we end up with a facet that is no longer subscribed
to controller state changes for re-render.

Technically could have done it with "unsubscribe only if the host is
also disconnected" kinda logic, but this should do just as well.

https://coveord.atlassian.net/browse/KIT-3517