docs(headless): typedoc plugin for coveo search box and theming (#4361)
https://coveord.atlassian.net/browse/KIT-3421

This is the first of a number of PRs to generate Headless documentation
using https://typedoc.org/
Here, I create a typedoc plugin to add an Atomic search box to the
generated website, and theme it.

I suggest merging to a typedoc-release-branch rather than to master
because the tags are in a messy state at the moment. I have enough to
work on the plugin, but that's it. My next task will be to go through
them to clean up and generate the typedoc website will all the content
we want. Avoiding master will also help to avoid conflicts with the v3
release. Typedoc can wait until v3.

You can preview by:
- checking out my branch
- running `npm install`
- running `npm run build:typedoc` from the headless package
- opening `packages/headless/docs/index.html`