fix(quantic): expandable searchbox not collapsing back to normal height after clearing  (#4379)
[SFINT-5700](https://coveord.atlassian.net/browse/SFINT-5700)

### ISSUE:


https://github.com/user-attachments/assets/92119536-8fd3-4d8e-8e48-3f6c7c27019c

When `clearInput` was called upon pressing the clear icon, it was
sending an input value change event with an empty value, focusing on the
input and then, if it was a textarea (expandable), it would call the
`adjustTextAreaHeight` function.

This `adjustTextAreaHeight` function did the following: 

`    this.input.style.height = this.input.scrollHeight + 'px';
`

[ScrollHeight Doc
](https://developer.mozilla.org/en-US/docs/Web/API/Element/scrollHeight)


> Btw this was called twice here (once when we called `clearInput` and
once when we called `onFocus` which was called in the `clearInput` so I
removed the unnecessary call.

The issue was that after focusing on the input, the
`this.input.scrollHeight` would still be equal to the height of the
content before clearing since the content itself was not wiped. It would
therefore adjust the height of the textarea with the previous content
height even though the value of the input was afterwards ''.

### SOLUTION:

By setting `this.input.value = ''` when clearing, we insure that
whenever the user clears by clicking the clear icon, it resets the
height of the textarea to its default height of 48px.

### DEMO:


https://github.com/user-attachments/assets/df2cc433-954f-4f95-9140-084e77c887e6




[SFINT-5700]:
https://coveord.atlassian.net/browse/SFINT-5700?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ