feat(generatedAnswerAnalyticsAction): stream id (#4405)
SVCC-4021

# Managing the answer api flow rga components custom events

## With the answer api flow
The answer id will be sent to the custom event

![image](https://github.com/user-attachments/assets/6aee4ff8-7501-423b-b3f7-d42da8ec3310)


## With the search Api flow
The search api stream ID will be sent with the custom event

![image](https://github.com/user-attachments/assets/d6aecf4a-5743-47ec-99d4-f39504c2334a)

## Description

It was decided that we should send a distinct key in the custom event
when using the answer api stream id. So I modified the selector to
return the `answerAPIEnabled` is boolean that will help determine what
key to use when sending the event. Not sure if this could have been done
in a better way.

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>