chore: update tsconfigs to match node version  (#4402)
https://coveord.atlassian.net/browse/KIT-2611

Tried to follow this for Node 20
https://github.com/microsoft/TypeScript/wiki/Node-Target-Mapping

Successful for most projects. Sometimes had to use modeResolution:
bundler and mode: es2022 instead of node16/node16. Not that big of a
deal. The important part is the "target". Updated the target for all
projects except stencil projects where it caused rollup issues (atomic &
atomic-hosted-page)