chore(atomic): fix tsconfig resolution for the e2e folders (#4462)
https://coveord.atlassian.net/browse/KIT-3604

In order to fix this problem, we have to tell the TS server to use
`"moduleResolution":"Bundler" & "module": "ES2022"` for the e2e/
folders. They are currently excluded in the atomic root tsconfig which
is why we get the error here.
<img width="1261" alt="image"
src="https://github.com/user-attachments/assets/a1956eda-a88b-4eb6-925d-c5d18cd8227f">

This is the solution I came up with, we would now have a tsconfig for
the IDE, a tsconfig for the stencil compilation, a tsconfig for the
storybook compilation and a tsconfig for playwright eslint rules 😅 . If
anyone has a better idea I am very open to it.