fix(answerApi): custom event (#4478)
# V2

This PR backports https://github.com/coveo/ui-kit/pull/4405 into the V2
branch.

## This is a fix

Since we should not backport features to the V2 branch, I will defend
that this PR is indeed a fix for the whole answer api feature on V2.
Without this. The whole feature is unusable for customers to test the
answerAPI properly. And will furthermore hinder the momentum we could
have to incentive them to move to v3 with the rest of the feature we
will add to the answerAPI in the next sprints.

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>