ci(atomic): run only affected E2E tests in atomic (#4484)
This pull request aims to optimize the E2E tests by running only the
affected tests.

Identifying only the affected tests by listing the changed files can be
tricky in the case of E2E testing. However, since we are testing
components in isolation, we can use this approach along with TypeScript
AST (Abstract Syntax Tree) to build a tree representation of the atomic
code structure. This allows us to create a dependency map between each
E2E test file and all the source files that might affect them.

### Rules
* If a change is detected in a source file that is a dependency (direct
or indirect) of an E2E test, the dependent test file will be added to
the list of tests to run.
* If a change is detected in an E2E test file (or any test-related
file), it will be added to the list of tests to run.
* If the changed file has no effect on E2E tests (e.g., updating the
README), no tests will be executed.
* If changes are detected in a different package (e.g., /headless), all
tests will be executed.
* If an error occurs in the script (e.g., no associated E2E test for a
change made in an atomic component), a warning will be thrown and all
tests will be executed.

 
### Shard reduction
I also adjusted the number of shards based on the number of results. 1
shard per test suite (24 shard max)

https://coveord.atlassian.net/browse/KIT-3634

---------

Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>