feat(atomic): added logic to send analytics when opening user actions (#4512)
## [SFINT-5757](https://coveord.atlassian.net/browse/SFINT-5757)

In this PR:

- adding the logic log analytics when opening the user actions timeline
feature in the insight panel.
- adding the property `excludedCustomActions` in the Atomic Insight User
Actions components and passing to the Headless controller to allow the
atomic user actions component to benefit from the already existing
feature in the Headless user actions controller that allows filtering
out unwanted events.


Demo showing the analytics being sent with `legacy` analytics:



https://github.com/user-attachments/assets/d66437d7-131f-4b42-bfbb-11ad0a6162ec



[SFINT-5757]:
https://coveord.atlassian.net/browse/SFINT-5757?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ