test(quantic): fixed quantic e2e after salesforce 25 winter release (#4559)
## [SFINT-5769](https://coveord.atlassian.net/browse/SFINT-5769)

- Fixed E2E tests for the refine modal content and the timeframe facet.
- Timeframe facet tests failed because of changes that came up with the
new SF winter 25 release, the way of displaying errors on the timeframe
inputs was changed by SF and caused the failure of some of our tests.
- The refine modal content tests failed because we were trying to select
a a standard facet value after applying a timeframe facet filter, the
issue was that with this timeframe filter no facet value of the standard
facet was found so the facet was not displayed, to fix this I added a
command to clear all filter before testing the standard facet.

[SFINT-5769]:
https://coveord.atlassian.net/browse/SFINT-5769?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ