fix(atomic): allow atomic loader to be deployed to the CDN (#4568)
https://coveord.atlassian.net/browse/KIT-3666

This PR allows the `defineCustomElements()` function to be deployed to
the CDN by simply adding it into the `/dist` folder instead of it's
custom folder.

It causes a problem in atomic-react where the `defineCustomElements`
import becomes `"import { defineCustomElements } from
'@coveo/atomic/dist/loader';"`.
That import is invalid because of the added `exports` field in atomic
V3. To fix this, we need to change the import back to how it was. I
created a script to do this. It changes it back to `"import {
defineCustomElements } from '@coveo/atomic/loader';"`.