fix(quantic): fixed sf proxy issue with excludedCustomActions property (#4566)
## [SFINT-5778](https://coveord.atlassian.net/browse/SFINT-5778)

The following issue is being logged in the Service console when passing
an array of string as the excludedCustomActions property to the user
actions toggle component:

<img width="502" alt="Screenshot 2024-10-21 at 3 34 20 PM"
src="https://github.com/user-attachments/assets/605a9eb8-8c36-43d4-a245-47d6962029d8">

This issue was not impacting the feature but it was painting the browser
dev console in red.
The manipulation added in this PR avoids it.


[SFINT-5778]:
https://coveord.atlassian.net/browse/SFINT-5778?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ