feat(quantic): adjusted logic to support custom tabs in the quantic tab bar component (#4619)
## [SFINT-4920](https://coveord.atlassian.net/browse/SFINT-4920)

Adjusted the logic in the Quantic Tab Bar component to allow it work
with custom tab component easily.

### Example:

instead of using the QuanticTab component to display my tabs, I created
a custom component called : `customTab`, I'm using this in my search
page as the following:

```
<c-quantic-search-interface>
   <c-quantic-tab-bar>
           <c-custom-tab  label="All" engine-id={engineId}></c-custom-tab>
           <c-custom-tab label="Articles" expression="@sfkbid" engine-id={engineId}></c-custom-tab>
   </c-quantic-tab-bar>
</c-quantic-search-interface>
```

I notice that my tabs are not being responsive, in smaller screens the
tabs overflow from my search page, that's because the QuanticTabBar is
not doing its job which is to always based on the width of the screen
and the width of the tabs do some calculation and decide which tabs are
the one that should be displayed and which should be displayed in a show
more drop down.
This is occurring because previously the Quantic Tab Bar components only
looks at the QuanticBar elements and nothing else.

 ### The solution

the Quantic Tab Bar components now looks at the QuanticBar elements and
every element that has a `data-role` set to the value `role`.

my code will become:
```
<c-quantic-search-interface>
   <c-quantic-tab-bar>
           <c-custom-tab  label="All" engine-id={engineId} data-role="tab"></c-custom-tab>
           <c-custom-tab label="Articles" expression="@sfkbid" engine-id={engineId} data-role="tab"></c-custom-tab>
   </c-quantic-tab-bar>
</c-quantic-search-interface>
```

and now my custom tabs are handled correctly by the Quantic Tab Bar
component!



https://github.com/user-attachments/assets/fb174203-06e3-4b3a-b84e-cfd1991f479a




[SFINT-4920]:
https://coveord.atlassian.net/browse/SFINT-4920?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ