fix(bueno): use URL constructor to parse url (#4644)
The previous URL validation disallowed localhost and 127.0.0.1-type
endpoints. Instead of using a regex to parse the URL, defer the parsing
to the URL constructor, which is more flexible.

To do so, I updated the TypeScript configuration to allow the use of
DOM-related types in the project.

This change came up as I was attempting to use `proxyBaseUrl` in
headless to call a locally-running instance of a service, which was
exposed on localhost. See
https://coveo.slack.com/archives/G016TA2G485/p1730933467640229