docs(headless): implement a better cart in headless-ssr-commerce sample (#4627)
https://coveord.atlassian.net/browse/KIT-3681

This PR implements a proper cart in our samples. 

To achieve this, I created server actions that simulate a server state
by using cookies (in memory state was not possible with both server
action and route handlers) in `@actions/cart.ts.` During the pages first
load we call the getCart server action to get the initialState.

I also created a util file in `@utils/cart.ts` where there are functions
for every cart use case. In each of those functions we update the
controller state and then call a server action in `@actions/cart.ts` to
update the server state of the cart.

---------

Co-authored-by: ylakhdar <ylakhdar@coveo.com>