feat(quantic): support of dependant facets added  (#4656)
## [SFINT-5773](https://coveord.atlassian.net/browse/SFINT-5773)

## Description

This PR introduces a new `dependsOn` public property to support
dependent facets functionality in our facet components. The `dependsOn`
property allows a facet to be conditionally displayed based on the
selection state of another "parent" facet. This feature provides greater
flexibility in controlling facet visibility, creating a more dynamic and
relevant filtering experience for users.

This feature in now supported for all types of facets: QuanticFacet,
QuanticCategoryFacet, QuanticNumericFacet, QuanticTimeframeFacet and
QuanticDateFacet.

## Property Structure

The `dependsOn` property accepts an object with the following fields:

- `parentFacetId` (string): The ID of the parent facet this facet
depends on.
- `expectedValue` (string): The value that must be selected in the
parent facet for this facet to be displayed.

## Example Usage

```javascript
  // exampleSearch.js
  ExampleDependency = {
    parentFacetId: 'filetype',
    expectedValue: 'txt'
  }
```

```html
  <!-- exampleSearch.html -->
  <c-quantic-category-facet
    with-search
    field="geographicalhierarchy"
    label="Country"
    depends-on={ExampleDependency}
    engine-id={engineId}
  ></c-quantic-category-facet> 
```

The category facet will only be displayed when the value `txt` is
selected in the `filetype` facet.



https://github.com/user-attachments/assets/e9e6423e-f8bd-40e0-a2cd-4c041885e94b






[SFINT-5773]:
https://coveord.atlassian.net/browse/SFINT-5773?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Louis Bompart <lbompart@coveo.com>