fix(quantic): fs-extra set as a dependency instead of a dev dependency to fix library instalation (#4669)
## [SFINT-5820](https://coveord.atlassian.net/browse/SFINT-5820)

The `fs-extra` npm library made as dependency instead of the dependency
to allow successful installation of the Quantic npm library.

The `fs-extra` library is used by both the pre-install and the
post-install scripts of the Quantic library and thus it needs to be
installed when the end user tries to install the @coveo/quantic npm
package.

[SFINT-5820]:
https://coveord.atlassian.net/browse/SFINT-5820?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ