refactor(headless, headless-react): no more generic engine in ssr-commerce types (#4657)
https://coveord.atlassian.net/browse/KIT-3730

The original ssr solution was made to receive multiple engine. For the
commerce use case, we had to duplicate a lot of the code since it was
not prepared to receive a solutionType. In that duplication, we copied
over the abstraction over the engine. It is not needed in ssr-commerce
since every ssr-commerce implementation will always use the same engine
which is SSRCommerceEngine.

---------

Co-authored-by: ylakhdar <ylakhdar@coveo.com>