feat(quantic): Fix the styling of the quantic sort component when inside refine modal (#4652)
[SFINT-4897](https://coveord.atlassian.net/browse/SFINT-4897)

## ISSUE:
The styling of the quanticSort component does not match what we do for
the filters and Atomic's. Changing this styling is not as simple
however. The `lightning-combobox` does not expose selectors to change
its styling such as height, border, etc.

## SOLUTION PROPOSED:

- Used variants for both "default" and "refineModal". When in variant
"refineModal", it will display the sort label differently and take the
whole width.
- Also deleted some of the deemed useless tests we had in quantic:
`should construct itself without throwing`

## LIMITATIONS:

- Although not perfect (ie: the height of the dropdown could be the same
as filters, shadowbox, etc.), this solution is simple and leverages SLDS
design system. So we know it works well and is accessible.

## POTENTIAL ALTERNATIVES:

- The alternative would be to create a custom dropdown component
whenever the quanticSort is in the refineModal. This is far more
complicated as we will no longer have access to the built-in
functionalities of the `lightning-combobox` for accessibility and so on.
It make the quanticSort component a bit more complex than it needs to
be.
- Another thing I tried was playing with the container's styling, but it
ended up being worse IMO.

## BEFORE:


![QuanticSort](https://github.com/user-attachments/assets/adbd8923-7273-408a-aa4f-380bdf38458c)


## AFTER:



https://github.com/user-attachments/assets/bad52617-370a-44b1-a754-9f9aaeb8cf4f



## TESTS:

<img width="572" alt="image"
src="https://github.com/user-attachments/assets/d75ef9c3-f68e-4626-a0a0-995b81560031">


[SFINT-4897]:
https://coveord.atlassian.net/browse/SFINT-4897?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ