fix(generatedAnswer): map state with answer api client (#4681)
SVCC-4353
https://coveord.atlassian.net/browse/SVCC-4353

## The problem

Upon interacting with a stream response, all headless selectors relying
on the `generatedAnswer` slice after a response fetched by the answerAPI
client aren't working properly because the state of the answerAPI client
is not mapped to the actual `generatedAnswer` slice.
e.g. the custom events

## The proposed fix

The answer API client is now dispatching what is relevant to the
`generatedAnswer` state. The selectors will then return the same value
as the generated answer controller state whenever it gets its answer
from the client getting its answer directly from reveal or from the
answer api.

Co-authored-by: Danny Gauthier <dgauthier@WKS-002169.corp.coveo.com>