test(quantic): first setup of playwright added in Quantic (#4597)
## [SFINT-5768](https://coveord.atlassian.net/browse/SFINT-5768)

### Intro
In this PR, we introduce the Playwright setup for the Quantic project,
allowing for automated E2E tests on Lightning Web Components (LWC) in
Salesforce. The Playwright structure here includes:

### Page Objects:
Page objects encapsulate component selectors and interactions, making
tests more modular and maintainable, in the proposed structure we have
common page objects and unique page objects.

The common page objects will be shared across all the Quantic E2E tests,
for example:

- `ConfigurationObject`: responsible of configuring the Quantic Example
page to initializes the test environment for a given test.
- `SearchObject`: responsible of performing a search and intercepting
the search requests in both the search use case and the insight use
case.
- `InsightObject`: responsible of performing the logic that is specific
to the insight use case, like for example waiting the initialization of
the Quantic Insight Interface.

A unique page object is an object that is specific to a given component,
like in this example the `PagerObject`

### Fixtures:
A Playwright fixture is a reusable setup that provides a consistent
environment for tests.

I created a base fixture called `quanticBase` that will be the base of
all the fixture we will create for each Quantic component, this base
fixture by default uses `ConfigurationObject`, `SearchObject` and
`InsightObject` page objects as these are always needed in any test.

Additionally each component will have his own fixtures that will extend
`quanticBase`, like for example in this example, the QuanticPager has
two fixtures `testSearch` and `testInsight`, two fixtures that provides
the setup to test the component in the search use case and the insight
use case respectively.

### Tests:

Following this structure, each Quantic component will now contain a new
folder called `e2e` that will contain the page object of the component,
the fixtures needed to test and the test spec.
Each component will have the following structure:
```
QuanticPager/
│
├── e2e/
│   ├── page-object.ts
│   ├── fixture.ts
│   └── quantic-pager.e2e.ts
├── __test__/
│   └── quantic-pager.unit.test.ts
│
├── QuanticPager.js
├── QuanticPager.html
├── QuanticPager.css
```
Each component will have the component logic, the unit tests and the E2E
in the same folder.

Jest Unit tests should focus on the following points: 

- Test component behaviour in isolation.
- Mock External Dependencies like the capabilities used from the
Headless library.
- Ensure the component renders the correct HTML based on its state and
input properties.
- Verify that component properties update correctly based on user
interactions or lifecycle events.
- Cover edge cases.

Playwright E2E tests should focus on the following points: 

- Test the essential paths that users take to achieve their goals using
a given component.
- Mimic actual user actions, such as clicking buttons, entering data and
navigation.
- Test interactions with external systems like our Coveo APIs and
analytics.



[SFINT-5768]:
https://coveord.atlassian.net/browse/SFINT-5768?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Frederic Beaudoin <fbeaudoin@coveo.com>
Co-authored-by: Simon Milord <simon.milord@gmail.com>
Co-authored-by: Etienne Rocheleau <erocheleau@coveo.com>