fix(quantic): issue with date filter and numeric filters when used with dependant facets (#4680)
## [SFINT-5822](https://coveord.atlassian.net/browse/SFINT-5822)

### The issue

Date filters and numeric filters in timeframe facets and numeric facets
respectively were not cleared when a parent facet that the numeric and
time frame facet depend on is cleared.

The expected behaviour is that when a parent facet is cleared, all the
facets that depend on it must also be cleared.


### The source of the issue

The issue was occurring because the numeric and date filters have
separate controllers, the `DateFilter` and `NumericFilter` controllers,
these controllers are different from the `DateFacet` and `NumericFacet`
controllers and these `DateFilter` and `NumericFilter` controllers
create facet objects with a `facet id` different from the one created by
`DateFacet` and `NumericFacet` controllers.

Since the facet id is different, the filters were not being take in
consideration by `Facet Condition Manager` and thus they were not
cleared when their parent facet is cleared.
This issue has been now fixed, by making the `Facet Condition Manager`
aware of both the facets and the filters for numeric and timeframe
facets.


### Before:


https://github.com/user-attachments/assets/e015f792-cbeb-404e-a15f-18a6a23aebc3


### After:


https://github.com/user-attachments/assets/110ea0e6-8663-490f-918f-c0285be86fe9




[SFINT-5822]:
https://coveord.atlassian.net/browse/SFINT-5822?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Simon Milord <simon.milord@gmail.com>