feat(quantic): answer config id property added in quantic generated answer component (#4660)
## [SFINT-5739](https://coveord.atlassian.net/browse/SFINT-5739)


### Description 
Added the answer configuration id property to the Quantic Generated
Answer component that will allow the component to start benefiting from
the Answer API by using the `generate` endpoint to generated answers.



https://github.com/user-attachments/assets/2b83920e-0098-452f-9f47-1ad5a1651181

#### Issue with the Quantic Generated Answer Content fixed:
I also fixed an issue with Quantic Generated Answer Content component,
previously this component was trying to load the marked dependencies
only on connected callback and when `answerContentFormat` is set to
`'text/markdown'`.
The problem with that is in the case where `answerContentFormat` is
initially set to `'text/plain'` and then is changed to
`'text/markdown'`, in this case the marked dependencies will never be
loaded because the connected callback will be re-executed.

Instead I updated the logic to make the component reactive to property
changes, now loading marked dependencies could be executed whenever the
value of `answerContentFormat` changes.



[SFINT-5739]:
https://coveord.atlassian.net/browse/SFINT-5739?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ