fix(headless): add globalThis shim to quanticUmd build to fix a bug with Locker Service (#4701)
Locker Service doesn't support `globalThis` but this was added to the
library coveo.analytics.js in the last version.

This meant all of Quantic components didn't load with Locker turned on.

This also made us realize that all of our Quantic test are only running
with LWS and not Locker, while the SFINT package tests are running on
Locker.

At least this way we cover both environments. Out of scope for this PR,
but in the future we will probably revert to testing on Locker with
Quantic too, and perhaps had smoke tests with LWS.

LWS being the "more recent" of the two, we expect less issues with LWS
than with Locker.


![image](https://github.com/user-attachments/assets/5e6c9eeb-919a-4f79-a2d0-ed7ad62c9e7b)