fix(headless): make `CommerceEngineDefinitionOptions` an optional generic type (#4703)
https://coveord.atlassian.net/browse/KIT-3713

The goal of this is for the implementer to not have to interact with the
`ControllerDefinitionsMap` & the `Controller` interface. These are
**internal** and bring more complexity to the implementer than anything
else. We can't simply remove that generic since it is used in
headless-react for inference but we can make it optional.

turns out this easier than I thought 😅. I did not think of simply making
it optional