feat(atomic): add experimental component for follow-up questions (#4729)
[SVCC-4341](https://coveord.atlassian.net/browse/SVCC-4341)

This PR adds a new experimental component providing follow-up questions
when a search query is performed. It relies on the Answer API to
retrieve the questions, so the `answer-configuration-id` attribute has
to be set on the component.


![follow-up-questions](https://github.com/user-attachments/assets/c3e2427e-e32f-411a-b9e1-f178de4b8d4e)

It can be added to a search interface this way:

```html
<atomic-follow-up-question-list answer-configuration-id="YOUR_ANSWER_CONFIG_ID"></atomic-follow-up-question-list>
```

`atomic-follow-up-question-list` subscribes to the `SearchBox` state
changes to detect when a new query is performed. When it happens, the
Answer API is called to retrieve follow-up questions for the current
query. The questions are then rendered by the
`follow-up-question-list-common` and `atomic-follow-up-question`
components.

At the experimental stage, follow-up questions are supported only for
the search use case. Eventually, the request would be performed from
headless and `follow-up-question-list-common` would be updated to make
it easy to support other use cases (i.e., insight and IPX).

[SVCC-4341]:
https://coveord.atlassian.net/browse/SVCC-4341?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: GitHub Actions Bot <>