fix(headless): capi 1587 facets are duplicated in capi request payload when using manual numeric facets (#4734)
[CAPI-1587](https://coveord.atlassian.net/browse/CAPI-1587)

Fix for a couple of issues in headless, as found in Barca with numerical
range facets (manual range facets in Headless)
- Remove duplicate facet's from the CAPI request
- Remove duplicate facet state from URL fragment
- Remove deselected numerical facet values from URL fragment

Curious I found a bug in the `actions.test.ts` file that meant the test
should have been failing due to the duplicate facet in CAPI request bug
but it passed. Fixing the bug now passes with the fixed code.

Current:

https://github.com/user-attachments/assets/9a075b2e-7dcb-4d65-b813-7f8ea8dc6129

Fix:

https://github.com/user-attachments/assets/0d78b0b2-dab0-400d-9483-23de0391912c




[CAPI-1587]:
https://coveord.atlassian.net/browse/CAPI-1587?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ