feat(github, quantic): create manual github workflow for Quantic E2E test setup (#4750)
[SFINT-5842](https://coveord.atlassian.net/browse/SFINT-5842)

## In this PR:

- We added a github workflow that can be triggered manually to
essentially create a scratch org, deploy the project and the Quantic
examples community and the e2e setup. When the workflow runs, it
generates the URL for the community with the setup done. This URL can be
used to test or write test locally.
- This allows QA for example to easily do the setup to write playwright
e2e tests and test components.

You can see the result of a successful run
[here](https://github.com/coveo/ui-kit/actions/runs/12166819370?pr=4750)


### DEMO SUMMARY:
<img width="1214" alt="image"
src="https://github.com/user-attachments/assets/4f69d901-4828-464d-8356-dd53f00b37a8">





[SFINT-5842]:
https://coveord.atlassian.net/browse/SFINT-5842?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ